refactor: move util to root directory
